ml-service/requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2

# Financial Data
yfinance==0.2.18
alpha-vantage==2.3.1
pandas-datareader==0.10.0

# Portfolio Optimization
cvxpy==1.3.2
pypfopt==1.5.5

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1
transformers==4.33.2

# Web Framework
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0
jupyter==1.0.0
